Rename editor header file to follow project's file structure
relates to #6845
