+++ VisualStudio/fheroes2/sources.props
@@ -322,9 +322,9 @@
     <ClInclude Include="src\fheroes2\dialog\dialog_selectitems.h" />
     <ClInclude Include="src\fheroes2\dialog\dialog_selectscenario.h" />
     <ClInclude Include="src\fheroes2\dialog\dialog_system_options.h" />
-    <ClInclude Include="src\fheroes2\editor\editor.h" />
     <ClInclude Include="src\fheroes2\editor\editor_interface.h" />
     <ClInclude Include="src\fheroes2\editor\editor_interface_panel.h" />
+    <ClInclude Include="src\fheroes2\editor\editor_mainmenu.h" />
     <ClInclude Include="src\fheroes2\editor\editor_map_specs_window.h" />
     <ClInclude Include="src\fheroes2\editor\editor_object_popup_window.h" />
     <ClInclude Include="src\fheroes2\editor\editor_castle_details_window.h" />
