+++ src/fheroes2/editor/editor_mainmenu.h
